fix: wrap Login page in the Suspense boundary
